requirements.txt
python-dateutil== 2.9.0.post0
requests==2.32.3
validator_collection==1.5.0
reportlab==4.3.1
#standard library -- re, csv, io, datetime, argparse, sys, pickle, os
